voice-publisher/requirements.txt
functions-framework==3.5.0
google-cloud-pubsub==2.19.0
google-generativeai==0.8.0
